Improvesments to make AdobeAnimate CreateJS files work better
I added code to Edit JS to try to fix conflicts between the production iframe in the block editor and AdobeAnimate assuming a global context. It was fixed for modern Animate files, but will still break in the editor for older Animate file. This is a problem with how AdobeAnimate does their rendering, NOT this createjs block